Fix bugs when following or unfollowing all record types
