+++ tests/integration/search/constants.py
@@ -1,0 +1,4 @@
+ENDPOINT_SEARCH_LOCATION_AND_RECORD_TYPE = "/search/search-location-and-record-type"
+TEST_STATE = "Pennsylvania"
+TEST_COUNTY = "Allegheny"
+TEST_LOCALITY = "Pittsburgh"
