+++ tests/integration/search/search_test_setup.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass
+
+from tests.helper_scripts.helper_classes.TestDataCreatorFlask import (
+    TestDataCreatorFlask,
+)
+from tests.helper_scripts.helper_classes.TestUserSetup import TestUserSetup
+
+
+@dataclass
+class SearchTestSetup:
+    tdc: TestDataCreatorFlask
+    location_id: int
+    tus: TestUserSetup
