+++ tests/integration/search/tests/follow/test_record_types.py
@@ -1,0 +1,4 @@
+def test_search_record_types():
+    pass
+
+    # Test functionality with record types
